requirements.txt
Flask
flask-restx
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn